setAbsoluteHeadingTarget no longer attempts to correct X and Y position
setAbsoluteHeadingTarget now uses the current pose's X and Y positions
as the X and Y targets to ensure that the subsequent call to
setAbsoluteTarget does not find any error along those axis, and thereby
turns off the X and Y PID controllers during the turn.

This ensures that the turn is not interrupted due to spurious error
along the X or Y axis while turning in place.